AutoFreestyler: support a generic comment syntax
This CL introduces the ability for comments in our examples to have any
key/value pairs, whilst also supporting the existing syntax. This means
we can define comments for examples either as:

```
<!--
  Why is this text not centered?
  # Because it has text-align: left; applied
-->
```

But now also:

```
<!--
  Prompt: Why is this text not centered?
  Explanation: Because it has text-align: left; applied
-->
```

The resulting object will be:

```
{
  prompt: 'Why is this...',
  explanation: 'Because...'
}
```

(In the old syntax, we assume lines with no prefix are prompts, and
lines with a `#` are part of the explanation).

Additionally, this CL also supports any random `A:B` key pairs, and also
ensures that multiple lines are supported. These will be used to provide
more context for other examples.

Bug: 383120286
Change-Id: Ib12c94c74d6045d845aad2edfb38d1dfb5166fec
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6171178
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Reviewed-by: Ergün Erdoğmuş <ergunsh@chromium.org>