RPP: fix text appearing on layout shift track
This bug was introduced by crrev.com/c/6172797 which mistakently caused
the layout shift track to have the "Layout shift cluster" text drawn on
to it.

The reason we didn't notice this before is because the layout shifts
track was trying to add the "Layout shift" title to each diamond, but
because the diamonds are not wide enough, no text got drawn onto it. But
that is not true of the clusters.

To fix this I reintroduced the `titleForEvent` call in the appender, but
make it always return an empty string, as we never want to title any of
the events on the track itself.

Fixed: 390093886
Change-Id: Ic77ca03dd94fce5ab040aac0d2254f6b0bb6b738
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6169939
Reviewed-by: Andres Olivares <andoli@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>