[AIAssistance] Simplify input shadow animation based on scroll
Previously, there were 2 things:
* We were setting the visibility depending on the scrollable
status of the container.
* We were updating the opacity of the shadow based on the
scroll status of the container.

We don't need to do these 2 things and we can go with
only setting the opacity when the animation is active
and setting it to `0` for the cases when the animation is
not active.

Bug: 373832540
Change-Id: I12f6eac31ba7636faa3c2dce33989bfaf7c53992
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6174730
Commit-Queue: Samiya Caur <samiyac@chromium.org>
Auto-Submit: Ergün Erdoğmuş <ergunsh@chromium.org>
Reviewed-by: Samiya Caur <samiyac@chromium.org>