Parse console.timeStamp payload
The payload now contains the new params the API was called with,
according to go/cpq:console-timestamp. This CL constructs the track data
with this data.

Timestamps for which a duration can be computed (because they have a start and an end) but that do not have an associated track name are stored separately. These events can be added to the timings track, I'd do that in a follow up CL.

Bug: 390155241
Change-Id: Icc5103e3e22849aabbd0dc5643c871c193084c77
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6172724
Commit-Queue: Andres Olivares <andoli@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>