[ui] Bring back dedicated `toolbar.css`.
The current styling for `<devtools-toolbar>` is difficult to comprehend,
because the global styles peek into deep DOM trees, and there's little
to no isolation. While it's beneficial to have the toolbar items in the
light DOM, it still makes sense to encapsulate the toolbar specific
styles in a shadow DOM. This is a first step to move the styling for the
toolbar itself into a separate CSS file again.

Drive-by-fix: Also add support for the `hidden` attribute.

Bug: 388445687
Change-Id: I4a74df36cf2bfdd21e65ed9067fd84a8f8fdf415
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6172348
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Auto-Submit: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>