RPP: determine if event can be used for AI Assistance
In crrev.com/c/6178276 we added the ability for the AI Assistance panel
to have the current timeline event updated as the user clicks around to
bring it in line with the other AI panels (e.g. elements). So rather
than have to use the context menu "Ask AI" button, you can also click
around, and then open the panel yourself.

However, one side-effect of this is that we started trying to construct
AI Call Trees for events that don't make sense; for example layout
shifts. In fixing that I ended up going down a bit of a rabbit hole; I
also discovered that no where had we really coded what the criteria was
for an event to be a valid event to pass into the AI Assistance panel.

If the AICallTree generation fails, it also breaks the panel and does
not let the user select an event, so we need to be careful about only
trying it on events that are legitimate.

This CL therefore tries to fix all of that by:

1. Ensuring that the method that creates AICallTrees returns `null` if
   the chosen event is invalid or if anything goes wrong. This ensures
   we do not break the panel and we can use this to see if an event is
   valid for AI assistance or not by calling it.
2. Update the custom context menu code to distinguish the possible
   filtering actions (collapsing children, etc) with the ability for an
   event to be used by the AI Assistance.
3. (Minor) cache the code that generates a list of threads for a given
   trace to avoid recreating it a lot, given that it doesn't change
   per-trace.

Fixed: 390507733, 390570738
Change-Id: I780a80674eef92521c2dc3aa87d681c667b47dce
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6182182
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Andres Olivares <andoli@chromium.org>