[source-map] Refactor away 'sourceMapToSourceList' weak map
The 'sourceMapToSourceList' maps from source map objects to a list of
sourceURLs which that source map contains. This is also useful when
dealing with index source maps.

Unfortunately this doesn't play nicely when resolving "sourceIndices"
from mappings, since you'd first have to resolve the raw source map
object, before you could resolve the full index.

This CL cleans this up by getting rid of the weak map alltogether:
We store the 'SourceInfo' objects in a plain array, indexed by the
"sourceIndex". For index maps, we simply concatenate all the sources
in the order of the "child source maps".

To preserve the existing behavior (and fast access), we also map
the 'sourceURL' to SourceInfo object in a separate map. If a source
map contains multiple 'sources' entries that resolve to the same
URL, then the first one wins. Same as the old implementation.

R=kimanh@chromium.org

Bug: 390337576
Change-Id: I2ae7c6b3d0667b98390085eb949cdee52375ef7f
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6180299
Reviewed-by: Kim-Anh Tran <kimanh@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>