Do not reset trace data when custom tracks setting changes
When the checkbox for the setting is clicked, we are calling
FlameChartDataProvider::reset, which as of [1] resets the trace data.
This is not the initially intended behavior as we only want to reset the
UI data. To fix this we simply call clearTimelineDataCache instead of
reset.

[1] https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6098009
Fixed: 391039190
Change-Id: I9987f00e6bfc2abfcd3d9c649fdc2f2b52b7f37e
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6185442
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Auto-Submit: Andres Olivares <andoli@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>