[BezierSwatch] Append text child when setBezierText is called
Prior to crrev.com/c/6113686, the text element was appended
as a child to the created legacy custom component.

However, in the  CL mentioned above, we removed legacy
custom components and migrated them to use `customElements.define`.

One drawback of this is, `document.createElement` cannot create
elements and their children at the same time. So, we were using
`connectedCallback` to append the child.

This subtly caused a bug: bezier swatches in the styles tab
did not have enough horizontal space from the start. The reason
is, PropertyRenderer checks whether to add a space between different
elements or not using `requiresSpace` method. This method decides
on this by checking the `textContent`s of the nodes.

Since the nodes are not connected to the DOM at that point, BezierSwatch
does not have the text element as the child. So, its `textContent` was
returning null.

This CL, updates `setBezierText` method to append the text element
if it is not already appended to the BezierSwatch element.

Fixed: 391149444
Change-Id: I930952c6335af3e3f576acb21782449cb02c5016
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6185437
Auto-Submit: Ergün Erdoğmuş <ergunsh@chromium.org>
Commit-Queue: Ergün Erdoğmuş <ergunsh@chromium.org>
Commit-Queue: Philip Pfaffe <pfaffe@chromium.org>
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>