Name Synthetic Layout Shifts as "SyntheticLayoutShift"
The root cause of this work is a bug where the user can select a layout
shift on the main thread, which lead me to discover that we named
`SyntheticLayoutShift` as `LayoutShift` which caused some confusion in
the code as in most places we work with synthetic layout shifts.

This CL fixes that and updates the code to match. We also mark "raw"
layout shifts as invisible as we only ever want to render synthetic
layout shifts.

This CL also removes a test for the layout shift UI from
TimelineUIUtils; we use the new layout shift component to render these,
so the test was obsolete.

Bug: 390578412
Change-Id: I025d7d4fb9ca9959539c63e63b7874a381c681de
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6189459
Reviewed-by: Andres Olivares <andoli@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>