Also monkey-patch `Document.adoptedStyleSheets`.
Previously we only monkey-patched the `ShadowRoot.adoptedStyleSheets`
accessor pair, but we need the same for `Document.adoptedStyleSheets`,
as we are also "sharing" constructed style sheets on the document level.

This also refactors the code to cache per `CSSStyleSheet` constructor,
instead of per `Document` instance, to make it easier to reason about.

Bug: 388445687
Change-Id: Ib0e4aec05f7fa3b0f7d8ea4851cd60c0ec2468bc
Fixed: 391716947
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6191869
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>