[sources] Remove constant re-creation of 'children' array
The NavigatorTreeNode in the 'Sources' panel stashes the children in a
Map internally but hands out an array for it's 'children' accessor.
The implementation spreads the Map#values into an array.

When debugging web apps with thousands of files that are deeply nested,
this becomes noticeable and a performance trace where we spend 25
seconds executing JS in DevTools, we spend ~500ms just spreading child
NavigatorTreeNodes.

This CL fixes this by maintaining both a map and an array for a tree
nodes' children. We also extract a new helper method 'updateId' so
we can use private fields for both the map and the array.

R=pfaffe@chromium.org

Bug: None
Change-Id: I42d1a7f6934dc05babbb055aaf9d19780004c759
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6190894
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>