Simplify handling of custom track setting change in Perf panel
Toggling the setting used to call a bunch of methods from
TimelineFlameChartView to rebuild the timeline to include/exclude
custom tracks. Thus it was hard to keep track of the functions
that needed to be called to ensure all UI items are reset and rebuilt
properly. For this reason, overlays weren't being rebuilt after the
setting was toggled.

This CL aims to simplify this by extracting the methods that build the
data in TimelineFlameChartView into a dedicated `refresh` method, which
is called when the trace model is set and when the setting is toggled.

Drive-by: make TimelinePanel more unit test friendly by accepting a trace model on instantiation and update old screenshot tests.

Fixed: 391328289
Change-Id: If9b89a36183821f28f7e8c8199eae84cc4d979ba
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6179454
Commit-Queue: Andres Olivares <andoli@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>