Use <devtools-new-datagrid> in the TrustTokenView
As this is using <devtools-button> in one of the cells and we rely on cloneNode to populate cells, override cloneNode in <devtools-button> to properly copy properties and event listeners. Not very happy about it, but this is probably a lesser evil.

The proper solution would be to reflect nodes instead of cloning them, but it requires slot assignment working for the indirect children (which is not supported), or moving nodes in some way, which would make observing changes very complicated.

Bug: 390346490
Change-Id: I48082008ef7a5adf0e915c00f711d9588614e8f9
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6190454
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Danil Somsikov <dsv@chromium.org>
Auto-Submit: Danil Somsikov <dsv@chromium.org>