RPP: fix 1/3P data table having first row selected by default
Turns out we have some logic to auto select rows. I don't know if we
need it, but it was a deliberate change for a11y reasons, so I don't
want to remove that functionality. So I've added a flag such that we can
disable it for this table.

Bug: 383567961
Change-Id: I7ab73e5f95e33cdf825476ad69a5897397f52772
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6198201
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Adriana Ixba <aixba@chromium.org>
Commit-Queue: Adriana Ixba <aixba@chromium.org>