[RPP] Don't outline insight related events if just from range overlay
When an insight is active, any range overlays it has marks all the
events that are relevant, and so should not be dimmed. Combined with the
outline treatment, that can be rather noisy given the number of events
involved.

We want to reduce how much outline we do, and only do it for events that
are sure to be of interest (and so may need the outline to even be
noticed).

So now the draw batching code for FlameChart, and the new dimmers, take
this into account by using an optional array of "what indices should be
outlined".

Batching code was refactored to support that we now need to batch by
outline mode, not just color.

Fixed: 392883608
Change-Id: If1f0a0c32bdccbf1bc33598b86c737399607e39b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6197543
Commit-Queue: Connor Clark <cjamcl@chromium.org>
Reviewed-by: Paul Irish <paulirish@chromium.org>