[RPP] Reduce draw time for dimming/outline in flame graph
A large portion of every frame's draw time (when zooming/moving the
trace window) was spent deciding if each entry should be dimmed or
outlined. To speed this up, cache the dim/outline state in an
Uint8Array.

Thanks to it being contiguous and only being used as an index, this
sped up this part of the drawing code by ~5x.

Bug: 40278532
Change-Id: I6f4f761071ba5c66b4c659c480964b6a41bec8ad
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6209089
Commit-Queue: Connor Clark <cjamcl@chromium.org>
Reviewed-by: Paul Irish <paulirish@chromium.org>