Make the controls tool the privacy panel default landing view
If there is no saved setting for the last selected view on the
privacy and security panel, the default should be the controls tool.

This change also fixes the focus behavior so that the sidebar will
color the last selected element with the selected color (blue) when
switching to the panel. Before, it would color with the hover (grey)

Bug:b/391604674
Change-Id: Ibffad102191b21abdd935c7c278f2ee250f7705d
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6199183
Reviewed-by: Shuran Huang <shuuran@chromium.org>
Reviewed-by: Paul Irish <paulirish@chromium.org>
Commit-Queue: Paul Irish <paulirish@chromium.org>