Parse trace ids from samples in trace CPU profiles
As per go/chrome-devtools:better-js-flamecharts, this CL does the
frontend bits to extract the trace id from samples and use it to obtain
the stack trace of the trace event with the same trace id. This stack
trace is used in the consolidation of tracing and sampling data that
creates the JS calls in the thread flamecharts.

Bug: 390155857
Change-Id: I6c21eb9a0b8ecf61422c249d6ebe9fa68091c3eb
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6218850
Commit-Queue: Andres Olivares <andoli@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>