Fix security sidebar memory behavior
Since we are no longer defaulting to the SecurityMainView, it can be
removed from the view function to let the sidebar set the
visible/main view. Without this, the view function was foribly
setting the SecurityMainView as the visible/main view when the panel
was constructed.

Bug: b/391604674
Change-Id: Ieb6658b430db53f76079866a51c5d516f3df3835
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6226579
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Commit-Queue: Joshua Thomas <masnoble@chromium.org>
Reviewed-by: Shuran Huang <shuuran@chromium.org>