[RPP 3P table] "Unhover" tree row after bottom up button click
This fixes a bug where after clicking the bottom-up button,
the 3P checkbox remains "disabled". This is because dimming on the
tree row hover remains after the click. So, it still thinks
that we are currently dimming on row hover of the 3P table.

So, we send an event to reset this

Bug:393955677
Change-Id: Id69129ca02674138404ab9ac7c936d4eb86cdcaf
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6228898
Auto-Submit: Adriana Ixba <aixba@chromium.org>
Reviewed-by: Paul Irish <paulirish@chromium.org>
Commit-Queue: Paul Irish <paulirish@chromium.org>