Fix card usage and add test
The card usage was changed in: http://crrev.com/c/6216268. This
caused issues for the CookieControlsView. This CL implements the
proper usage and adds a small e2e test to make sure changes to the
card will be caught when component changes are first made.

Bug: 394081331
Fixed: 394204700
Change-Id: Iac32e67fb8d0a4dfec5c5d439cf7fff62af34306
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6226581
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Reviewed-by: Shuran Huang <shuuran@chromium.org>
Commit-Queue: Joshua Thomas <masnoble@chromium.org>