[css] show crossed-out values on hover
In the styles pane we cross out values that are unused, e.g., fallbacks
in a var(), colors in light-dark(), values in min(), and so forth.
Crossed out values are hard to read, so we remove the line-through when
hovering the values.

Bug: None
Change-Id: I7843a88ec72ed890f27c863b1387e2651c310132
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6243916
Auto-Submit: Philip Pfaffe <pfaffe@chromium.org>
Reviewed-by: Ergün Erdoğmuş <ergunsh@chromium.org>
Commit-Queue: Ergün Erdoğmuş <ergunsh@chromium.org>