Reveal the CookieReportView by selecting sidebar element
This fixes the issue where the someone may click a console message
that reveals the CookieReportView, but the wrong sidebar element is
highlighted. By making the revealer select the sidebar element
directly, we allow the sidebar to manage showing the the
CookieReportView as it would with a normal user selection.

Bug:394934461
Change-Id: Iada064f42db049ce7e9a0f40c0438886684e891e
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6241738
Reviewed-by: Shuran Huang <shuuran@chromium.org>
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Commit-Queue: Shuran Huang <shuuran@chromium.org>