AI: Add base Performance Insights agent
This CL adds the foundations for the new Perf Insights agent. It is not
usable yet as we never set the right `flavor` when the user selects an
Insight, so the agent is never triggered.

Additionally to enable it you must enable the feature param:

```
--enable-features="DevToolsAiAssistancePerformanceAgent:insights_enabled/true"
```

Which is false by default.

Bug: 394552594
Change-Id: If5771962a78653c9150307b81134a8a81f59737f
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6244423
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>