Enable DeferRendererTasksAfterInput
With DeferRendererTasksAfterInput shipping in Chrome by default, there
is no guarantee that that all triggered by the CDP input (for example,
Input.dispatchMouseEvent) will be processed before the next task (for
example, Runtime.evaluate) runs. This makes the tests that do not wait
for the expected page state to become more flaky if the input task was
delayed.

Tests like these should be treated as other flaky tests and either fixed
or disabled for investigation.

Bug: 361078921
Change-Id: I4002b9014e0af1576d5d319c67663b2087a86007
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6226408
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>