Apply UI eng vision to the KeyValueStorageItemsView
To make it work, implement refresh callback in the data grid web component add make VBox understand element as the first parameter.

This could be further improved, but removing the view output and propagate the preview value through this HTML template, but this is left for later.

Bug: 394287937
Change-Id: Ic92aaeace4bf90610a46e7882c43736f5451a578
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6249842
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Danil Somsikov <dsv@chromium.org>