Improve extraction of stack trace for extension events
The stack trace is computed from the profile call that made the call to
the extension API. Before https://crrev.com/6252783, we could not
determine with certainty which profile call called the API for extension
entries added with performance.measure, since we didn't have a trace
event representing the API call. With the new event parsed in that CL we
can determine which call it is by looking at the event's hierarchy and
finding the first profile call ancestor.

Bug: 395845218
Change-Id: If9ab0e69c0737f5dbf7e49bc13543544a5b35c51
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6252828
Commit-Queue: Andres Olivares <andoli@chromium.org>
Reviewed-by: Nancy Li <nancyly@chromium.org>