[Animations] Fix animation ui not getting rendered
After the empty state update (crrev.com/c/6254938), we started
not displaying the animations container. However, it was used to
calculate the width of the timeline.

This CL updates the logic for calculating the timeline width
and calculates it based on the full width of the widget instead
looking at the animations container.

Bug: none
Change-Id: I1fe608527f73e3068030401b997719c36ce68834
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6262908
Commit-Queue: Ergün Erdoğmuş <ergunsh@chromium.org>
Auto-Submit: Ergün Erdoğmuş <ergunsh@chromium.org>
Reviewed-by: Kim-Anh Tran <kimanh@chromium.org>