[project_settings] Add "workspace.uuid" field.
Adapt the implementation to the latest design changes, and require a
"uuid" in addition to the "root". Also change the validation of the
`devtools.json` to require "workspace" to be either absent or an object
that contains both "root" and "uuid" string properties.

Doc: https://goo.gle/devtools-json-design
Bug: 395037775, 395562934
Change-Id: Ia73568152abe7150db3823d210470e791c6a7b90
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6264208
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>
Auto-Submit: Benedikt Meurer <bmeurer@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>