[Perf panel] Prefer profile calls over trace events when sorting
If timestamp and duration are equal between a trace event and a profile
call, the profile call should be put first when sorting in ascending
order. This is because when the timesamp of a profile call is exactly
the same as a trace event's, it's likely because the SamplesIntegrator
sets the profile call's ts so that it ends up parenting the trace event
in the flame chart. Without this extra check in the comparator, the
precedence meant by the samples integrator isn't certain to happen.

Bug: 389056697
Change-Id: I8c3754782397ab640888ee3fa5fa3e91772e6aef
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6264603
Reviewed-by: Alina Varkki <alinavarkki@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>