AI: add Perf Insights formatter
This CL continues the work to lay the foundations of the Performance
Insights agent. Note that as of this CL the agent remains not usable and
is not expected to provide much value.

- Added `PerformanceInsightFormatter` class to format insight details,
  links, and descriptions.
- Updated `PerformanceInsightsAgent` to use the new formatter.
- Added `isLCPPhases` type guard to `LCPPhases.ts`. This allows us to
  take a generic insight and narrow the type down.

Bug: 394552594
Change-Id: I8b514bdb6d6ca404eca4f168a30254c0fba7b0e8
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6276133
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>