[SRI Message Signatures] Handle errors delivered to devtools.
This CL implements the devtools side of the integration for SRI
Message Signature errors that occur during parsing or validation in
the network service. It copies over the protocol changes from the
Chromium side of the patch, and fleshes out the issue implementation
and messaging.

1. [Chromium] https://crrev.com/c/6163804
2. [Devtools frontend] https://crrev.com/c/6239271 [You are here]

Bug: 381044049
Change-Id: I94df831bcaacce8b37a60f880d5e01dc7e374db9
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6239271
Commit-Queue: Mike West <mkwst@chromium.org>
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Reviewed-by: Simon Zünd <szuend@chromium.org>