RPP: add AIQueries to look up relevant data
This CL adds the foundations for an `AIQueries` class that can, for a
given Insight, return the relevant parts of data. Right now we are
starting with network events.

For a given Insight, the querying code will:

1. Calculate the relevant bounds for the insight. This will be from the
   associated navigation (or start of trace) to the next navigation (or
   end of trace). For some insights (currently only "LCP by phase",
   although this will be expanded), we scope the bounds specifically to
   fit the insight.
2. Go through the network requests and find the ones that fit within the
   calculated bounds.

Note that the code to pass this data into the LLM is not written yet and
will be done in a follow-up.

Bug: 394552594
Change-Id: Id5f4f551535bdc33cc2d07a6552e1e583faa5e79
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6276829
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Nancy Li <nancyly@chromium.org>
Commit-Queue: Nancy Li <nancyly@chromium.org>