Make DOM breakpoints empty state only show a one-liner
With https://crrev.com/c/6152575 we updated the empty states for the
DOM breakpoints view. However, DOM breakpoints are shown both in
sources and in the elements panel. For the sources panel, we only
want to show a one liner.

Drive-by: Slightly changing the header text to fit in with the remaining sidebar empty state one-liners
Screenshot: https://imgur.com/a/dbRtGdN
Fixed: 397254690
Change-Id: I5bb70e7462c9e0b1f0863b3ecaf8cfc73e2e4fd8
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6277065
Reviewed-by: Kateryna Prokopenko <kprokopenko@chromium.org>
Commit-Queue: Kim-Anh Tran <kimanh@chromium.org>