AI: define `getNetworkActivity` for Perf Insights agent
This CL allows the LLM to call the function to get the network activity
for a given insight.

In the design doc I explored using time ranges, but so far I am going
for an approach where we query based on the active insight. I think this
reduces the chance of the AI incorrectly passing invalid or incorrect
times, but we can explore and experiment.

Bug: 394552594
Change-Id: I2d5c054391d5802ce46f5c149ee7ce194630fa68
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6276941
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Nikolay Vitkov <nvitkov@chromium.org>
Commit-Queue: Nikolay Vitkov <nvitkov@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>