RPP: Add AI Query to fetch main thread activity
Note: this CL does not give the LLM access to it, that will be a
follow-up.

This CL provides a function that, for a given insight, can return
relevant main thread activity. We might want to expand this in the
future or explore how much context window this might use, but this is a
good starting point.

Bug: 394552594
Change-Id: Iea4eeb8bd87e6d2e04acc84c86536bb3f69a90e9
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6276942
Reviewed-by: Nancy Li <nancyly@chromium.org>
Commit-Queue: Nancy Li <nancyly@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>