[Animations] Fix animations freezing screencast view
When an animation is captured, we start a screencast
for taking its screenshots for preview and after
finishing taking screenshots, we stop the screencast.

However, if there already exists a screencast (i.e.
if the user is inspecting a remote page using `chrome://inspect`),
after the animations finished taking screenshots, they
stop the screencast for `ScreencastView` too causing the
view to freeze.

This CL:
* Adds ability to stack screencasts in screen capture model.
When it receives a new `startScreencast`, it adds the
existing screencast to a stack and serves to the newly added
screencast.
* Then, when the newly added screencast is stopped, it resumes
the previous screencast from the stack.

Fixed: 395838062
Change-Id: I7ee6df0cbb4e9f7927cb1cb250ceb69e78355a77
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6276143
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Commit-Queue: Ergün Erdoğmuş <ergunsh@chromium.org>