[console] Wrap Error.stack in a span/div when standard formatting fails
This CL wraps the linkified Error.stack DocumentFragment in a span or
div element when the "tryFormatAsError" method fails to extract and
render an Error object.

With this CL, we properly separate the different Errors in a
Error.cause chain where we fail to parse and format some of the
Errors.

R=bmeurer@chromium.org

Fixed: 395489589
Change-Id: Id0e4d610281e4e1f5cf7fef9d75a595ec545a6c5
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6276558
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>