[RPP Observations] Handle prerendered pages correctly
Clicking a link to a prerendered page could trip up the live metrics
model if the frame from the original page is removed before live metrics
handles the interaction event.

This CL skips the step to check if a frame is the primary frame if it
comes from the already known "good" execution context.

Bug: 394918742
Change-Id: I8773d0771690a55a583afaf4ed3d97f4058346ee
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6284568
Reviewed-by: Connor Clark <cjamcl@chromium.org>
Commit-Queue: Adam Raine <asraine@chromium.org>