[styles] make color-mix keyboard interactable
This change allows selecting color-mix swatches with the keyboard by
using the left/right arrows when a style rules is selected. When a
color-mix is focused, alt+down activates its popover.

To get the focusing right (focus only the icon and not the entire
function call), the color-mix text is moved out of the icon's DOM
structure.

Bug: 396311936
Change-Id: Ibd3b095961beaf97a6538579a605b2955ec076ce
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6286012
Reviewed-by: Ergün Erdoğmuş <ergunsh@chromium.org>
Commit-Queue: Philip Pfaffe <pfaffe@chromium.org>