Remove one of two toolbar comboboxes. We had ToolbarCombobox and ToolbarComboBox
The former is a base class of ToolbarMenuButton, and otherwise is only used in the NetworkLogView, with a use case that better fits ToolbarMenuButton. 

Therefore, keeping ToolbarComboBox and removing the ToolbarCombobox, but folding it into the ToolbarMenuButton and migrating the NetworkLogView to it.

Had to update NetworkLogView test as these were reliant on the soft context menu DOM structure. 

Bug: 301364727
Change-Id: I2b2d6791335d58b3b6d7150e3dff47aec47e0f65
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6291332
Auto-Submit: Danil Somsikov <dsv@chromium.org>
Commit-Queue: Danil Somsikov <dsv@chromium.org>
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>