Revert disabling device emulation for chrome:// pages
The original change to disable device emulation for chrome:// pages
introduces a race condition that results in inconsistent behavior.

When the user navigates from a chrome:// page to a non chrome:// page,
the navigation event causese the emulation to be enabled. However,
the page could already have been partially loaded when the emulation
setting takes effect. Depending on when this happens, the emulated
result could look different.

Since this approach is fundamentally unsound, let's revert this change
and instead accept that we'll have issue reports of chrome:// pages
that are not compatible with device emulation.

Fixed: 388411685
Change-Id: I690b5ed89c20602c46e590f89367dc5f775fd8ac
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6298051
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>
Auto-Submit: Yang Guo <yangguo@chromium.org>
Commit-Queue: Philip Pfaffe <pfaffe@chromium.org>