[AiAssistance] Make ChatView more presentational
Previously, it was deciding on "what" text to render
based on the parameters. Now, for the text that's changing
based on the business logic; it accepts an input that is
supplied by the AiAssistancePanel.

Drive-by:
* Previously, after `aidaAvailability` is set in panel
construction, it was getting updated after a call to
`checkAccessPreconditions` when it's shown. I have
stubbed that as well and added a helper function to
stub it in tests.

Bug: 394005781
Change-Id: Ie111d397dee7c5cd117c651271c2e57453898160
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6298050
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Ergün Erdoğmuş <ergunsh@chromium.org>