CSS value tracing for arithmetic functions
This CL adds a new UI for tracing the evaluation of CSS values. This
first iteration adds the UI on hover of custom property declarations in
the styles tab. Supported arithmetic functions are min, max, clamp, and
calc initially, more may be added later as needed. This CL does not
support substitutions of var()s yet, those will be added in a followup.

Bug: 396080529
Change-Id: I16b66bbfb66dbfa01c7a9630a28992450a052e46
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6269777
Reviewed-by: Ergün Erdoğmuş <ergunsh@chromium.org>
Commit-Queue: Philip Pfaffe <pfaffe@chromium.org>